Bail out of checkParseSafety early on too-long data
Files with huge XMP blobs were throwing warnings in XML parsing
due to an internal limit in libxml2 of 1 MiB for URIs. Since
the data is parsed through a data: URI for safety, this limit
was triggering on affected files.

This should reject the metadata on affected files cleanly, without
throwing warnings.

Bug: T242966
Change-Id: I12461b38a0de577b82b3ad54e36ab01abd3dc800