Add DigitalSourceType field to the extraction list
This is what XMP uses to mark a file as being generated by AI.
Most AI generators do not use XMP to mark their files, but
given how hot button an issue this is, I think its worth supporting
existing metadata standards in this area.

The field is defined at
https://iptc.org/std/photometadata/specification/IPTC-PhotoMetadata#digital-source-type
It's value is a url starting with
http://cv.iptc.org/newscodes/digitalsourcetype/

This will also need a corresponding commit in MW to display the
url in a more human friendly form.

Bug: T405138
Change-Id: I8ff4b4170b033d6b1f22aada91bf0601d887ce29